Randomized Artemis and error-feedback (#8)
Implement Randomized Artemis (RArtemis)
Implement the error-feedback mechanism but only for downlink compression (and thus only for algorithms with bidirectional compression. I am still wondering if there is not a bug inside?
Re-implement the CompressionModel class and set up new compressor operators.
Re-organize pre-defined parameters to ease their usage and customization.
Experiments on both RArtemis and EF.
Reformat the way real datasets are preprocessed (now done in the code and no more in the notebooks).